Refactored base command files
