+++ src/commands/game/queue/empty-queue.ts
@@ -1,5 +1,5 @@
 import { Message } from "discord.js";
-import { BaseCommandClass } from "../../../BaseCommand";
+import { BaseCommandClass } from "../../../BaseCommand/BaseCommand";
 import { print, deleteDiscMessage } from "../../../tools/f";
 import { QueueAction } from "./queue";
