+++ src/commands/stats/teammates.ts
@@ -3,8 +3,8 @@ import * as f from '../../tools/f';
 import { printMessage } from '../../bot';
 import { GameModesStandard, getGameModes, getModeChosen } from '../../game/gameModes';
 import { Message } from 'discord.js';
-import { BaseCommandClass } from '../../BaseCommand';
-import { MatchMode } from '../../BaseCommandTypes';
+import { BaseCommandClass } from '../../BaseCommand/BaseCommand';
+import { MatchMode } from '../../BaseCommand/BaseCommandTypes';
 
 export const commands = ['teammates', 'friends'];
